requirements.txt
# Core dependencies
openai>=1.0.0
aiohttp>=3.10.11
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.7.0
requests>=2.32.2
urllib3>=2.2.2
orjson>=3.9.15
cryptography>=44.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Code quality
black>=24.3.0
mypy>=1.0.0
pylint>=3.0.0
ruff>=0.1.0

# Security
safety>=2.3.0
bandit>=1.7.0

# Utilities
cachetools>=5.3.0
rich>=13.0.0
pyyaml>=6.0.0

# Monitoring
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Additional dependencies
aiohttp-retry>=2.8.3
